Improve handling of html
Still not 100% satisfactory. If we edit a note with the html editor, even if we cancel our changes, the serialized html can be different from the intial html (with some meaningless whitespace added or removed here and there). But handling this can be very tricky, since two html strings can be different and rendered the same way.